+++ common/src/main/java/com/devnemo/nemos/inventory/sorting/Constants.java
@@ -6,6 +6,8 @@ import org.slf4j.LoggerFactory;
 public class Constants {
 
 	public static final String MOD_ID = "nemos_inventory_sorting";
+	public static final String IRON_CHEST_MOD_ID = "ironchest";
+	public static final String NEMOS_BACKPACKS_MOD_ID = "nemos_backpacks";
 	public static final Logger LOG = LoggerFactory.getLogger(MOD_ID);
 
 	public static final int MAX_MERGING_CYCLES = 1000;