+++ common/src/main/java/com/devnemo/nemos/inventory/sorting/helper/ButtonTypeMapping.java
@@ -1,0 +1,6 @@
+package com.devnemo.nemos.inventory.sorting.helper;
+
+import com.devnemo.nemos.inventory.sorting.factory.ButtonCreator;
+
+public record ButtonTypeMapping(String componentName, ButtonCreator factory, int defaultYOffset, boolean isInventoryButton) {
+}